Threadpool + ParallelAdapters
Add Threadpool implementation

These are threadpool implementations for parallel adapters that are used within the RSR algorithm. I decided to make this separate so I can easily swap out implementations which can be useful during profiling as multithreading can sometimes make results more ambiguous. In addition, we cannot use the built-in parallel algorithms. Not only is C++20 missing a good number of algorithms that we are interested in using (or have variants that are undesirable to use in some way), but one of the compilers that we wish to support (Apple Clang) does not support execution policies at all. Even when execution policies are supported, the C++ standard places no requirements on implementors to actually execute relevant algorithms in parallel when parallel execution is requested.

The ThreadPool class seems to still contain a subtle race condition where the pool can deadlock if the given closure finishes really quickly. I initially wanted to use this design because it amortizes thread creation costs but I might end up just removing it as our algorithms take long enough that thread creation is typically completely negligible.

Add Parallel Adapters

Adds concepts and functions for parallel algorithms.

These add parallel algorithms comparable to ones available in the algorithm header. Note that execution policy is still not supported on Apple Clang right now (2025-07-22) which just so happens to be the compiler Andreas uses.

This library is much stricter than the algorithms one in the sense that we require the input and output ranges to be contiguous and have constant time random access. While there are more generic ways of writing parallel algorithms, cache locality can lose us more performance (>10x) than what we gain thanks to multithreading.

Also note that the names follow functional programming concepts rather than the C++ std names (map, filter instead of transform, erase_if) since the choice of some of the std names were picked to either avoid collisions with existing portions of the library or follow some specific C++ idiom which we don't necessarily care about.

It is also worth noting that there is quite a bit of duplication here to express some more complex adapters that would normally be composed in more functional code. This is because we need to block to force computations to complete. Allowing composition without blocking would require a Rust style library where computation is indirectly described in terms of traits and an explicit collect is added to the end to finish the computation. It is possible to write this in C++ but it is overly complex and most of the computation that we are doing can be described in terms of a simple map-fold.

Note that I might also turn IExecutor into a template parameter to allow the compiler to inline one more level although I suspect the impact of thi son performance is negligible.

Add ThreadPool tests

ParallelAdapters:

Use proper cache line size

ParallelAdapters: fix FilterFunctor concept

ParallelAdapters: 64 is the cache line size for me and you ♫