RawObj + Parsers
Add RawObj class and parsers

There are several ways of parsing obj files in GEL right now. None of which fit my particular needs since these are meant to be used to feed TriMesh or Manifold. I need a parser that just provides the raw data and the ad hoc ones that are spread all across the demo files won't do.

Note that we will almost certainly want to move this to the test directory eventually. It would be preferable if outside users would not see yet another obj parser and parser tend to be security landmines.

Add Obj parser tests

Not comprehensive tests for the parser. Likely to be removed soon.

RawObj now supports faces

RawObj: Fix a face parsing bug, properly handle no normals case

RawObj: Due to MSVC lacking the relevant overloads wrt from_chars and string_view, we have to use this awkward juggling of data() and length() instead of begin() and end().

RawObj handles an edge case

There are some cases where the obj file has no faces but still has normals which are equal in number to the vertices. We can then map each normal directly to one vertex.