Add RawObj class and parsers
There are several ways of parsing obj files in GEL right now. None of which fit my particular needs since these are meant to be used to feed TriMesh or Manifold. I need a parser that just provides the raw data and the ad hoc ones that are spread all across the demo files won't do.

Note that we will almost certainly want to move this to the test directory eventually. It would be preferable if outside users would not see yet another obj parser and parser tend to be security landmines.