Make most of the CGLA API constexpr
This might seem like a fairly elaborate change, but it is actually rather mundane. Constexpr is C++'s way of indicating that a function (or a value) can be evaluated at compile time. This has two preconditions. The first is whatever operation performed has no side effects (i.e. it is a pure function, with whatever definition you would use for pure in C++), the second is it does not have undefined behavior. The entire definition also has to be visible (although this should fall under the second rule somewhat unintuitively).

Ultimately I do not expect this to have any impact on performance (in fact, if it does, that means the optimizer was doing a very poor job before). But it does allow CGLA functions to be used in static_assert and other constexpr functions and it forces the optimizer to constant fold certain things that it was not guaranteed to before.