Add AssociativeContainers.h
Adds type definitions for types used in the parallel_hashmap.

See #102 for context. But due to the slow nature of the built in associative containers, I think that this is a highly beneficial addition.

Adds a mutable priority queue as an alternative to generational priority queues.

This class generalizes a priority queue by allowing key ordering relations to be stored indirectly

This means that keys can be removed arbitrarily and not just from the front. Existing keys can be updated by reinserting them to the queue.

Note that the memory footprint of this is substantially larger than an ordinary vector-based min/max heap. In addition, the internal queue is always kept sorted, meaning there is a potentially higher up-front cost to constructing this. So make sure you really need the mutability or really benefit from it when using this over a generational priority queue.

Add missing assertion import and size() and empty()