Add SparseGraph inspired by the graph implementations inside less_slow.cpp
I am not too happy about this one. The graph implementation that is most likely to outperform the built-in one would have to use flat_hash_set/flat_hash_map from the phmap library. The problem however is that those containers only return a single element when given equal_range which we want to abuse to store edges with the power of heterogeneous keys. This of course also means that the built in std::unordered_set and std::unordered_map are actually a multi set and a multi map which is horrifying.

I will however consider keeping this implementation since I think it has some ergonomic benefits over the AMGraph and we might perhaps figure out some way to actually exploit its niche advantages.

(It is also possible that the real reason for the performance deficit is merely that this collection only really performs great for sparse graphs and our connectivity is just a tad too high in RSR)