RsR2: Many improvements
Make SimpGraph no longer derived. (I'm not so sure about this change and might revert it since we have to expose the inner part anyway).

Instead of using AttribVec, use std::vector in SimpGraph and RSGraph. This results in a major performance improvement.

Revert SimpGraph and RSGraph to use AMGraph. The SparseGraph idea wasn't bad but we have too many edges for it to be anywhere near performant.

Move get_neighbor_info and maintain_edge_loop inside RSGraph. That's where they make sense.

Convert a few pairs to dedicated structs to avoid potential memcpy fails.

Convert a lot of values to references to avoid copying. As it turns out this was one of our biggest performance losses because each Vec3 is 24 bytes and certain functions are copying quite a few of them. In this case, inlining becomes undesirable because the indirect memory access will nearly always beat copying so much data to the stack over and over.

Add some experimental multithreading to edge connection.