ArithVec: get rid of dumb algorithms for elementary operations.
This change seems to have a performance impact of up to 20%, which is simply astonishing. It is likely that GCC (and other compilers), simply do not understand what std::transform and similar functions do at a high level and completely fail to optimize these to efficient SIMD ops. The end result is a lot of useless runtime checked tiny loops that run useless integer ops and waste branch prediction slots.